Refactor PsgParamType to use enum class, update PsgMidi and PsgList for improved type safety and clarity, simplified MidiPsg tests

+++ src/plugins/uZX/aychip/AYPlugin.h
@@ -60,7 +60,7 @@ struct ParamAttachment {
         referTo(id, n, def, u);
         choices.clear();
         for (int i = 0; i < ch.size(); ++i) {
-            choices.push_back({static_cast<Type>(i), ch[i]});
+            choices.push_back({static_cast<Type>(static_cast<size_t>(i)), ch[i]});
         }
     }
